+++ app/(auth)/(guest-only)/layout.tsx
@@ -15,7 +15,7 @@ import {userQueryKey} from '../../queries/common/queryKeys';
 // 3. 프로필 드롭다운에서 마이페이지 버튼 클릭
 // 4. token이 없음으로 login 페이지로 redirect
 // 5. 이 경우 GNB에는 로그인 버튼이 떠야하는데 회원정보가 그대로 남아있게 된다.
-function AuthCleaner({children}: {children: React.ReactNode}) {
+function GuestOnlyLayout({children}: {children: React.ReactNode}) {
   const queryClient = useQueryClient();
 
   useEffect(() => {
@@ -27,4 +27,4 @@ function AuthCleaner({children}: {children: React.ReactNode}) {
   return children;
 }
 
-export default AuthCleaner;
+export default GuestOnlyLayout;
